# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuerySage API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type DBConfigDbType = typeof DBConfigDbType[keyof typeof DBConfigDbType];


export const DBConfigDbType = {
  postgresql: 'postgresql',
  mysql: 'mysql',
} as const;

export interface DBConfig {
  db_type: DBConfigDbType;
  host: string;
  port: number;
  database: string;
  username: string;
  password: string;
}

export type AnalyzeInputDbType = typeof AnalyzeInputDbType[keyof typeof AnalyzeInputDbType];


export const AnalyzeInputDbType = {
  postgresql: 'postgresql',
  mysql: 'mysql',
} as const;

export interface AnalyzeInput {
  query: string;
  db_type: AnalyzeInputDbType;
  db_config?: DBConfig;
  /** @nullable */
  manual_schema?: string | null;
  /** @nullable */
  explain_output?: string | null;
}

export type BottleneckSeverity = typeof BottleneckSeverity[keyof typeof BottleneckSeverity];


export const BottleneckSeverity = {
  HIGH: 'HIGH',
  MEDIUM: 'MEDIUM',
  LOW: 'LOW',
} as const;

export interface Bottleneck {
  type: string;
  /** @nullable */
  table?: string | null;
  description: string;
  severity: BottleneckSeverity;
}

export interface IndexSuggestion {
  statement: string;
  reason: string;
}

export interface ExplainNode {
  node_type: string;
  cost?: string;
  rows?: string;
  execution_time?: string;
  details?: string[];
  children?: ExplainNode[];
}

export interface OptimizationResult {
  /** @nullable */
  id?: number | null;
  original_query: string;
  optimized_query: string;
  explanation: string;
  bottlenecks: Bottleneck[];
  suggested_indexes: (string | IndexSuggestion)[];
  estimated_improvement: string;
  execution_plan_summary: string;
  db_type: string;
  /** @nullable */
  created_at?: string | null;
  /** @nullable */
  query_complexity_score?: number | null;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface HistoryEntry {
  id: number;
  original_query: string;
  optimized_query: string;
  explanation: string;
  bottlenecks: Bottleneck[];
  suggested_indexes: (string | IndexSuggestion)[];
  estimated_improvement: string;
  execution_plan_summary: string;
  db_type: string;
  created_at: string;
  user_id: string;
  query_complexity_score?: number;
  /** @nullable */
  share_id?: string | null;
  chat_history?: ChatMessage[];
}

export interface CreateShareInput {
  history_id: number;
}

export type UserStatsTopBottleneckTypesItem = {
  type: string;
  count: number;
};

export interface UserStats {
  total_optimizations: number;
  postgresql_count: number;
  mysql_count: number;
  top_bottleneck_types: UserStatsTopBottleneckTypesItem[];
}

export interface ChatRequestBody {
  history_id: number;
  message: string;
}

export type CreateShare200 = {
  share_id: string;
};