# lib/api-zod/src/generated/types/analyzeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuerySage API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzeInputDbType } from './analyzeInputDbType';
import type { DBConfig } from './dBConfig';

export interface AnalyzeInput {
  query: string;
  db_type: AnalyzeInputDbType;
  db_config?: DBConfig;
  /** @nullable */
  manual_schema?: string | null;
  /** @nullable */
  explain_output?: string | null;
}