# lib/api-zod/src/generated/types/dBConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuerySage API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DBConfigDbType } from './dBConfigDbType';

export interface DBConfig {
  db_type: DBConfigDbType;
  host: string;
  port: number;
  database: string;
  username: string;
  password: string;
}