# lib/api-zod/src/generated/types/explainNode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuerySage API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExplainNode {
  node_type: string;
  cost?: string;
  rows?: string;
  execution_time?: string;
  details?: string[];
  children?: ExplainNode[];
}