# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuerySage API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Bottleneck } from './bottleneck';
import type { ChatMessage } from './chatMessage';
import type { IndexSuggestion } from './indexSuggestion';

export interface HistoryEntry {
  id: number;
  original_query: string;
  optimized_query: string;
  explanation: string;
  bottlenecks: Bottleneck[];
  suggested_indexes: (string | IndexSuggestion)[];
  estimated_improvement: string;
  execution_plan_summary: string;
  db_type: string;
  created_at: string;
  user_id: string;
  query_complexity_score?: number;
  /** @nullable */
  share_id?: string | null;
  chat_history?: ChatMessage[];
}