# lib/api-zod/src/generated/types/optimizationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuerySage API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Bottleneck } from './bottleneck';
import type { IndexSuggestion } from './indexSuggestion';

export interface OptimizationResult {
  /** @nullable */
  id?: number | null;
  original_query: string;
  optimized_query: string;
  explanation: string;
  bottlenecks: Bottleneck[];
  suggested_indexes: (string | IndexSuggestion)[];
  estimated_improvement: string;
  execution_plan_summary: string;
  db_type: string;
  /** @nullable */
  created_at?: string | null;
  /** @nullable */
  query_complexity_score?: number | null;
}